Fix Mender NPC and food check handling
Renamed "Mender" to "Merchant & Mender" so it targets the NPC correctly.

The food function was looping infinitely due to a missing text node element it expects to find, which isn't present with the /item command. Replaced this check with the logic that it attempts food 3 times and moves on instead.